Action filter for Model validation, exceptions filter. All commands are returning OperationResult<T> now

+++ CwkSocial.Application/UserProfiles/CommandHandlers/DeleteUserProfileCommandHandler.cs
@@ -1,11 +1,14 @@
-﻿using CwkSocial.Application.UserProfiles.Commands;
+﻿using CwkSocial.Application.Enums;
+using CwkSocial.Application.Models;
+using CwkSocial.Application.UserProfiles.Commands;
 using CwkSocial.DataAccess;
+using CwkSocial.Domain.Aggregate.UserProfileAggregate;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
 namespace CwkSocial.Application.UserProfiles.CommandHandlers;
 
-public class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfileCommand>
+public class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfileCommand, OperationResult<UserProfile>>
 {
     private readonly DataContext _context;
 
@@ -14,14 +17,26 @@ public class DeleteUserProfileCommandHandler : IRequestHandler<DeleteUserProfile
         _context = context;
     }
 
-    public async Task<Unit> Handle(DeleteUserProfileCommand request, CancellationToken cancellationToken)
+    public async Task<OperationResult<UserProfile>> Handle(DeleteUserProfileCommand request,
+        CancellationToken cancellationToken)
     {
         var userProfile = await _context.UserProfiles
             .FirstOrDefaultAsync(profile => profile.UserProfileId == request.UserProfileId);
 
+        var result = new OperationResult<UserProfile>();
+        if (userProfile is null)
+        {
+            result.isError = true;
+            var error = new Error
+                { Code = ErrorCode.NotFound, Message = $"No UserProfile with ID {request.UserProfileId}" };
+            result.Errors.Add(error);
+            return result;
+        }
+
         _context.UserProfiles.Remove(userProfile);
         await _context.SaveChangesAsync();
-        
-        return new Unit();
+
+        result.Payload = userProfile;
+        return result;
     }
 }
